Create AGI_Predictive_Calculator.py
AGI Predictive Calculator

This is a prototype of an AGI predictive calculator. It uses a TensorFlow model to predict the user's next calculation when they enter an expression. The model is trained on a dataset of expressions and their corresponding predictions. The predicted calculation is then suggested to the user.

This calculator can be used for a variety of purposes, such as:

Helping students learn math
Helping scientists with their research
Helping software developers with their work
Helping people with their everyday tasks
Features

Predicts the user's next calculation
Can be used for a variety of purposes
Usage

To use the calculator, simply enter an expression and press enter. The calculator will then evaluate the expression and print the result. The calculator will also suggest the next predicted calculation.

Example

Enter an expression: 2 + 2
The result is: 4
Suggested next calculation: 4 * 4
Contributing

If you would like to contribute to the development of this calculator, please feel free to fork the repository and submit a pull request.

Saving to GitHub